refactor(chromium): isolate Yandex protocol specifics into yandex.go
Move source overrides, AAD, intermediate-key derivation, and master-
password gate to yandex.go. Yandex / Opera extract functions remain
alongside defaults in extract_*.go for horizontal symmetry.
